Change project structure
Create package for deta base code and separate base module to few
modules.
