+++ deta_py/deta_base/__init__.py
@@ -1,0 +1,7 @@
+"""Deta Base module.
+
+Include classes and utilities for Deta Base API operations.
+
+See https://deta.space/docs/en/build/reference/deta-base
+for details about Deta Base API.
+"""
