+++ setup.cfg
@@ -13,7 +13,7 @@ per-file-ignores =
     # There are multiple fixtures, `assert`s, and subprocesses in tests:
     tests/*.py: S101, S105, S404, S603, S607, WPS110, WPS211, WPS226, WPS323, WPS442
     # items is an official name for Base object
-    deta_py/base.py: WPS110
+    deta_py/deta_base/*: WPS110
 
 [isort]
 profile = wemake
